requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
requests>=2.31.0
beautifulsoup4>=4.12.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
drf-yasg>=1.20.0
tenacity>=8.2.0 
groq